requirements.txt
langchain==0.0.311
pandas==2.1.1
python-dotenv==1.0.0
streamlit==1.27.2
openai==0.28.1
tabulate==0.9.0